Move unit tests to test directory
