+++ package.json
@@ -33,7 +33,7 @@
     "eslint": "^7.2.0",
     "eslint-config-hackreactor": "github:hackreactor-labs/eslint-config-hackreactor",
     "faker": "^4.1.0",
-    "jest": "^26.0.1",
+    "jest": "^26.1.0",
     "node-fetch": "^2.6.0",
     "nodemon": "^2.0.4",
     "pg-copy-streams": "^5.1.0",
